pyproject.toml
[project]
name = "PolyBot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.7",
    "dotenv>=0.9.9",
    "httpx>=0.28.0",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.2.1",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "rich>=15.0.0",
    "visidata>=3.3",
]
